+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes in this project will be documented in this file.
 ## [0.4] - 2020-09-10
 ### Added
 - add a text transformer implementation
+- implement text util to strip all chars that are no characters or digits
 
 ### Deprecated
 - TextNormalizerService - use the TextTransformer instead